remove epoch
